Fetch more than one hardcoded feed
We add the wonderful "IT hare on software" to the hardcoded lineup.

It brings new and exciting problems, like that its authors include their
novel-sized posts in the RSS description. We should probably just print
the post title in that case, rather than try to figure out how to
extract any kind of description from it.