Refactor project
Keeping everything in main.go is getting a bit cluttered, and the names
I have for things suck. I've wanted to try structuring a project along
the lines described in [1] for a while, so here is an attempt at that.

Another reason for the split is that I hope it'll make for easier
testing.

We now define a reader package, that contains Feed and Entry interfaces.
We implement those interfaces in an rss package that also fetches RSS
feeds over HTTP, and have the HTTP server in cmd/evil-rss-reader render
and serve the results.

I can see an argument for splitting the RSS fetching from its parsing
and so on, but I don't have a clear split in mind and we don't have to
do that now.

[1] https://medium.com/@benbjohnson/standard-package-layout-7cdbc8391fc1