Refactor feed provider away from parser
The rss package was in charge of:

    1. Fetch feed updates over HTTP
    2. Parse RSS feeds
    3. Implement the reader.Feed interface

Its job really should only be the second two, so we abstracted away the
code that fetches updates over HTTP, and added ones that also fetch
updates on the local file system or from a string. (The latter two we
can use while developing.) We can also profit from this when we
implement atom feed parsing.