Stop parsing feed times
It is our own Vietnam.

Yes, the Atom and RSS specs lay out how their timestamps should be
formatted. Also, given a sample of ~20 feeds, we've witnessed 3
common variations each on both the Atom and RSS formats, plus one
YYYY-MM-DD out of left field.

Just because you spec it, doesn't mean people follow the spec.

So screw it. We don't need to know the exact time something was
published anyway. The patch that follows this one will instead implement
a "Seen" policy, where we keep track of whether a feed's items changed
during an update, and mark it as having new items as long as we have yet
to view it.

I think that in the end, I like that approach better than the
date-parsing one, though I might just be saying that because by now I
hate humanity and its date formats.