Keep track of "seen" feeds
We now keep track of a hash of the last N items we read from a feed.
When the hash of the items changes during an update, we mark the feed as
having new items.

This introduces state that we need to keep track of between program
restarts, so we do so by serializing it to JSON on disk.