Treat HTML in Atom titles
The static site generator Hugo apparently creates item titles with HTML
and CDATA. I think we could deal with that by having item and feed
titles be template.HTML, but I'm hesitant to do that. This is mostly
because I think the RSS spec is not super clear on whether HTML is
allowed in titles. For example, I believe [sic] that

    A < B

written exactly like that is a valid title, but it would create
wonkyness if we tried to render that literally as HTML inside a title.

Instead we unescape HTML in titles by hand if we detect it, for Atom
feeds.