Ignore SSL certificates
This is a temporary hack. We will figure out how to properly deal with
SSL certificates for the next release.

Some blogs use self-signed SSL certificates, which the HTTP client
doesn't like. I don't want to assume that we can install SSL
certificates on the servers that host the reader, so we'll probably
end up letting users point to extra certificates that we'll use when
fetching feeds.