Create cmd directory for main package
I'd like to create a command-line version of the reader as well, and
keep it in the same directory tree, so I'm moving the existing reader to
its own directory.

This made the bindata unhappy and I wasn't in much of a mood to figure
out why, so the static data is just a string in the sources now.