Set HTTP timeout
The default GO HTTP client has a timeout of 30 seconds, which seems a
bit excessive:

    https://golang.org/src/net/http/transport.go?s=#L40

We set ResponseHeaderTimeout in the Transport, because the client
timeout includes time spent reading the response body. That takes quite
some time at the moment, because XML parsing is slow, and we don't want
to time out there.