Fix resource leak in ScatterGather
We never returned from the select statement in the goroutine that was
collecting respones and errors. The select statement would keep polling
on closed channels forever, and quickly tied up 100% of all our CPUs.
We can avoid this by just range-ing over the channels, so the goroutines
in which we do that return once the channels have been closed.