Adhere to XDG base dir spec
We now look for a config file according to these rules:

If defined, use the value from the -config command-line option.
Otherwise, check for a file called 'evil-feed-reader.yaml' in:

- The directory the binary is running in
- The directories specified in the XDG spec

Use the first of these that exists.

If none of these files exist, fall back to 'config.yaml' in the
directory the binary is running in.