Fetchers grow context
Having one feed take very long to respond to any kind of request made
all the feeds slow, and the whole program so slow that my nginx proxy
was timing me out. If we make our fetchers take a context, we can pass
that to the HTTP request we make, and time out like civilized people.